+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tidypress/config
 
+## 1.0.15
+
+### Patch Changes
+
+- Release alignment with `tidypress@1.0.15`.
+
 ## 1.0.12
 
 ### Patch Changes